Move new types to common point
`ObjectSpecType` and `UpdateToSpec` should not be mergebot-specific.
