+++ src/test/java/com/codicesoftware/plugins/hudson/model/ObjectSpecTest.java
@@ -1,6 +1,6 @@
 package com.codicesoftware.plugins.hudson.model;
 
-import com.codicesoftware.plugins.jenkins.mergebot.ObjectSpecType;
+import com.codicesoftware.plugins.jenkins.ObjectSpecType;
 import org.junit.Test;
 
 import static org.junit.Assert.assertEquals;
