+++ src/test/java/com/codicesoftware/plugins/jenkins/UpdateToSpecTest.java
@@ -1,4 +1,4 @@
-package com.codicesoftware.plugins.jenkins.mergebot;
+package com.codicesoftware.plugins.jenkins;
 
 import hudson.AbortException;
 import org.junit.Test;
